+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [Version 1.0.0](https://github.com/dataiku/dss-plugin-sendmail/releases/tag/v1.0.0) - Feature release - 2024-02
+
+- REMOVED SUPPORT FOR PYTHON 2.7 AND 3.5. The code env used must be in range python 3.6 - 3.11 inclusive
+
+- Features added
+  - templating capabilities - use JINJA style templating in the email body and subject line
+  - send as HTML
+  - ability to use DSS channels to send emails
+
 ## [Version 0.2.0](https://github.com/dataiku/dss-plugin-sendmail/releases/tag/v0.2.0) - Feature release - 2023-08
 
 - Added support for python versions 3.8, 3.9, 3.10 and 3.11 (DSS-12)